+++ lib/netstd/Thrift/Protocol/TProtocol.cs
@@ -85,18 +85,18 @@ namespace Thrift.Protocol
 
         protected void CheckReadBytesAvailable(TSet set)
         {
-            Transport.CheckReadBytesAvailable(set.Count * GetMinSerializedSize(set.ElementType));
+            Transport.CheckReadBytesAvailable((long)set.Count * GetMinSerializedSize(set.ElementType));
         }
 
         protected void CheckReadBytesAvailable(TList list)
         {
-            Transport.CheckReadBytesAvailable(list.Count * GetMinSerializedSize(list.ElementType));
+            Transport.CheckReadBytesAvailable((long)list.Count * GetMinSerializedSize(list.ElementType));
         }
 
         protected void CheckReadBytesAvailable(TMap map)
         {
             var elmSize = GetMinSerializedSize(map.KeyType) + GetMinSerializedSize(map.ValueType);
-            Transport.CheckReadBytesAvailable(map.Count * elmSize);
+            Transport.CheckReadBytesAvailable((long)map.Count * elmSize);
         }
 
         // Returns the minimum amount of bytes needed to store the smallest possible instance of TType.
